And tests and doc for `Duplicates`
